backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.2.1
email-validator==2.1.0.post1
python-multipart==0.0.6
pydantic-settings==2.1.0
email-validator>=2.0.0
bleach>=6.0.0

# Firebase & Database
firebase-admin==6.4.0
google-cloud-recaptcha-enterprise==1.25.0

# Machine Learning & Data Science
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0

# Computer Vision
opencv-python-headless==4.9.0.80
Pillow==10.2.0

# HTTP & Networking
requests==2.31.0
httpx==0.26.0
bleach==6.1.0

# Security
cryptography==42.0.0
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate Limiting
slowapi==0.1.9

# Monitoring & Logging
sentry-sdk==1.40.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
pylint==3.0.3
black==24.1.1
isort==5.13.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.6
